requirements.txt
Flask==2.1.0
Werkzeug==2.0.3
cryptography==36.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-admin==1.6.0
flask-login==0.6.0
flask-table==0.5.0
xlsxwriter==3.0.3
python-dotenv==0.19.2
psycopg2
gunicorn
